+++ planners-preprocess/src/scc.rs
@@ -1,16 +1,16 @@
 #[derive(Debug, Clone)]
 pub struct Scc {
-    graph: Vec<Vec<i32>>,
+    graph: Vec<Vec<usize>>,
     dfs_numbers: Vec<i32>,
     dfs_minima: Vec<i32>,
     stack_indices: Vec<i32>,
-    stack: Vec<i32>,
-    sccs: Vec<Vec<i32>>,
+    stack: Vec<usize>,
+    sccs: Vec<Vec<usize>>,
     current_dfs_number: i32,
 }
 
 impl Scc {
-    pub fn new(graph: Vec<Vec<i32>>) -> Self {
+    pub fn new(graph: Vec<Vec<usize>>) -> Self {
         Self {
             graph,
             dfs_numbers: Vec::new(),
@@ -22,7 +22,7 @@ impl Scc {
         }
     }
 
-    pub fn get_result(mut self) -> Vec<Vec<i32>> {
+    pub fn get_result(mut self) -> Vec<Vec<usize>> {
         let node_count = self.graph.len();
         self.dfs_numbers = vec![-1; node_count];
         self.dfs_minima = vec![-1; node_count];
@@ -32,7 +32,7 @@ impl Scc {
 
         for i in 0..node_count {
             if self.dfs_numbers[i] == -1 {
-                self.dfs(i as i32);
+                self.dfs(i);
             }
         }
 
@@ -40,10 +40,10 @@ impl Scc {
         self.sccs
     }
 
-    fn dfs(&mut self, vertex: i32) {
+    fn dfs(&mut self, vertex: usize) {
         let vertex_dfs_number = self.current_dfs_number;
         self.current_dfs_number += 1;
-        let v = vertex as usize;
+        let v = vertex;
         self.dfs_numbers[v] = vertex_dfs_number;
         self.dfs_minima[v] = vertex_dfs_number;
         self.stack_indices[v] = self.stack.len() as i32;
@@ -51,7 +51,7 @@ impl Scc {
 
         let successors = self.graph[v].clone();
         for succ in successors {
-            let succ_idx = succ as usize;
+            let succ_idx = succ;
             let succ_dfs_number = self.dfs_numbers[succ_idx];
             if succ_dfs_number == -1 {
                 self.dfs(succ);
@@ -64,11 +64,11 @@ impl Scc {
         if self.dfs_minima[v] == vertex_dfs_number {
             let stack_index = self.stack_indices[v] as usize;
             let num_stack_entries = self.stack.len();
-            let mut scc: Vec<i32> = Vec::new();
+            let mut scc: Vec<usize> = Vec::new();
             for i in stack_index..num_stack_entries {
                 let node = self.stack[i];
                 scc.push(node);
-                self.stack_indices[node as usize] = -1;
+                self.stack_indices[node] = -1;
             }
             self.stack.truncate(stack_index);
             self.sccs.push(scc);
